close issues # 11 close issues # 19

+++ bot/view/main.py
@@ -10,20 +10,23 @@ class View:
         self.chat_id = context.from_user.id
         self.cache = Cache()
 
-    async def print_message(self, text: str, kb) -> None:
+    async def print_message(self, text: str, kb, **kwarg) -> None:
         user = await self.cache.get_user(user_id=self.context.from_user.id)
 
-        try:
-            await self.context.bot.edit_message_text(chat_id=self.chat_id, message_id=user.active_msg_id, text=text,
-                                                     reply_markup=kb(user.active_msg_id))
-        except Exception as e:
-            print(e)
-            await self.delete_message(user.active_msg_id)
-            message_data = await self.context.bot.send_message(chat_id=self.chat_id, text=text,
-                                                               reply_markup=kb(user.active_msg_id + 1))
+        if kwarg:
+            print(kwarg)
+        else:
+            try:
+                await self.context.bot.edit_message_text(chat_id=self.chat_id, message_id=user.active_msg_id, text=text,
+                                                         reply_markup=kb(user.active_msg_id))
+            except Exception as e:
+                print(e)
+                await self.delete_message(user.active_msg_id)
+                message_data = await self.context.bot.send_message(chat_id=self.chat_id, text=text,
+                                                                   reply_markup=kb(user.active_msg_id + 1))
 
-            user.active_msg_id = message_data.message_id
-            await self.cache.update_user(user=user)
+                user.active_msg_id = message_data.message_id
+                await self.cache.update_user(user=user)
 
     async def delete_message(self, msg_id: int | None = None) -> None:
